build: move vale gradle config
